fix: adding requested changes
