+++ server/connection.ts
@@ -1,0 +1,20 @@
+import mysql from "mysql2";
+import "dotenv/config";
+
+const connection = mysql.createConnection({
+  host: process.env.DB_HOST,
+  user: process.env.DB_USER,
+  password: process.env.DB_PASSWORD,
+  database: process.env.DB_NAME,
+  charset: "utf8mb4",
+});
+
+connection.connect((err) => {
+  if (err) {
+    console.error(err);
+    return;
+  }
+  console.log("Successfully connected to the database.");
+});
+
+export default connection;
